moved components to compoenents folder/set up context api
